[FIX] remove circular imports
